+++ compiler/src/iree/compiler/Codegen/Common/GPU/test/BUILD.bazel
@@ -26,6 +26,8 @@ iree_lit_test_suite(
             "gpu_distribute_scf_for.mlir",
             "gpu_distribute_shared_memory.mlir",
             "gpu_generalize_named_ops.mlir",
+            "gpu_pack_to_instrinsics.mlir",
+            "gpu_fuse_and_hoist_forall.mlir",
             "gpu_greedily_distribute_to_threads.mlir",
             "gpu_infer_memory_space.mlir",
             "gpu_lower_to_ukernels.mlir",
