+++ compiler/src/iree/compiler/Codegen/Dialect/GPU/Transforms/BUILD.bazel
@@ -54,9 +54,7 @@ iree_compiler_cc_library(
         "CombineBarrierRegions.cpp",
         "ConcretizeMmaShapes.cpp",
         "DistributeMmaToLanes.cpp",
-        "FuseAndHoistParallelLoops.cpp",
         "LowerIREEGPUOps.cpp",
-        "PackToIntrinsics.cpp",
         "Passes.cpp",
         "Transforms.cpp",
         "UnrollToIntrinsics.cpp",
