+++ compiler/src/iree/compiler/Codegen/Dialect/GPU/Transforms/CMakeLists.txt
@@ -48,9 +48,7 @@ iree_cc_library(
     "CombineBarrierRegions.cpp"
     "ConcretizeMmaShapes.cpp"
     "DistributeMmaToLanes.cpp"
-    "FuseAndHoistParallelLoops.cpp"
     "LowerIREEGPUOps.cpp"
-    "PackToIntrinsics.cpp"
     "Passes.cpp"
     "Transforms.cpp"
     "UnrollToIntrinsics.cpp"
