+++ compiler/src/iree/compiler/Codegen/Dialect/GPU/Transforms/test/BUILD.bazel
@@ -21,8 +21,6 @@ iree_lit_test_suite(
             "combine_barrier_regions.mlir",
             "concretize_mma_shapes.mlir",
             "distribute_mma_to_lanes.mlir",
-            "fuse_and_hoist_forall.mlir",
-            "pack_to_intrinsics.mlir",
         ],
         include = ["*.mlir"],
     ),
